Nlpik2

Approved-by: Markus Giftthaler <markusgft@gmail.com>

+++ ct_optcon/include/ct/optcon/nloc/NLOCBackendST-impl.hpp
@@ -39,7 +39,7 @@ template <size_t STATE_DIM, size_t CONTROL_DIM, size_t P_DIM, size_t V_DIM, type
 void NLOCBackendST<STATE_DIM, CONTROL_DIM, P_DIM, V_DIM, SCALAR, CONTINUOUS>::computeLQApproximation(size_t firstIndex,
     size_t lastIndex)
 {
-    if (lastIndex == this->K_ - 1)
+    if (lastIndex == static_cast<size_t>(this->K_) - 1)
         this->initializeCostToGo();
 
     for (size_t k = firstIndex; k <= lastIndex; k++)
